+++ inst/bash/new_bash_script.sh
@@ -212,6 +212,3 @@ $CHMOD $DEFAULTSCRIPTRIGHT $OUTPUTPATH
 end_msg
 
 
-#' ## Latest Update
-#+ latest-update, eval=TRUE
-date