+++ NEWS.md
@@ -1,3 +1,8 @@
+## qgert 0.2.4
+
+* Adding new functionality of batch difference between directories
+* Improved spin_sh.R to include date of latest change.
+
 ## qgert 0.2.3 (Minor release)
 
 * Add comparison plots for deregression plots in GS